Add pub date to advanced search
