+++ app/components/trln_argon/advanced_search_form_component.rb
@@ -3,25 +3,6 @@ module TrlnArgon
     include RangeLimitHelper
     include ViewHelpers
 
-    # update range inputs to only allow numbers
-    def render_range_input(solr_field, type, input_label = nil, maxlength = 4)
-      type = type.to_s
-
-      if params['range'] && params['range'][solr_field] && params['range'][solr_field][type]
-        default = params['range'][solr_field][type]
-      end
-
-      html = label_tag("range[#{solr_field}][#{type}]", input_label, class: 'sr-only') if input_label.present?
-      html ||= ''.html_safe
-      html + text_field_tag("range[#{solr_field}][#{type}]",
-                            default,
-                            maxlength: maxlength,
-                            class: "form-control range_#{type}",
-                            pattern: '[0-9]+',
-                            oninvalid: "setCustomValidity('Publication Year should be a number')",
-                            oninput: "setCustomValidity('')")
-    end
-
     private
 
     # Override of https://github.com/projectblacklight/blacklight/blob/ce114b6c3709e09efcd2fa91429b84dd6f2ca08b/app/components/blacklight/advanced_search_form_component.rb#L39
