+++ app/components/trln_argon/advanced_search_range_limit_component.html.erb
@@ -1,0 +1,13 @@
+<!-- Fetch and display publication year range input if configured. -->
+<div class="form-group row advanced-search-facet">
+  <%= label_tag @facet_field.label.parameterize,
+    class: "col-12 col-sm-6 col-md-3 col-form-label font-weight-bold" do %>
+    <%= @facet_field.label %>
+  <% end %>
+  <div class="col-12 col-sm-6 col-md-9 range_limit">
+    <label for="range_publication_year_isort_begin" class="sr-only">Publication date range (starting year)</label>
+    <%= render_range_input(@facet_field.key, :begin) %> – 
+    <label for="range_publication_year_isort_end" class="sr-only">Publication date range (ending year)</label>
+    <%= render_range_input(@facet_field.key, :end) %>
+  </div>
+</div>
